% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile}
\alias{reconcile}
\title{Reconcile a data frame against an external data source}
\usage{
reconcile(
  data,
  endpoint,
  query_col,
  property_cols = NULL,
  limit = 10,
  matches_only = FALSE
)
}
\arguments{
\item{data}{A data frame of candidates for reconciliation, one per row}

\item{endpoint}{The URL of the reconciliation API endpoint, including API key
if required}

\item{query_col}{The name of the column to use for the main query}

\item{property_cols}{A vector of column names to use for additional
properties (optional)}

\item{limit}{The maximum number of reconciliation matches to return for each
candidate—defaults to 10}

\item{matches_only}{Whether to return reconciliation matches on their own,
without candidate data—defaults to "FALSE"}
}
\value{
The original data frame with reconciliation matches as additional
  columns, or a data frame of matches on their own if matches_only is
  "TRUE".
}
\description{
This is the primary function for matching data against a reconciliation
endpoint using the \href{https://reconciliation-api.github.io/specs/latest/}{Reconciliation Service API standard}.
}
\examples{
gw_companies <- tribble(
  ~name, ~jurisdiction_code, ~country_code,
  "Global Witness", "gb", "gb",
  "Global Witness", "be", "be",
  "Global Witness", "us_ca", "us")

gw_companies \%>\%
  reconcile(
    endpoint = "https://reconcile.opencorporates.com?api_token=TOKEN",
    query_col = "name",
    property_cols = c("jurisdiction_code", "country_code")) \%>\%
 filter(match_score >= 75)
}
