% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{widgets}
\alias{widgets}
\title{The output of a hierarchical forecast in a keys, index, outcome format.}
\format{
\subsection{\code{widgets}}{

A data frame with 20 rows and 5 columns:
\itemize{
\item \code{date} A date column.
\item \code{country} A grouping or keys column giving the forecast hierarchy.
\item \code{state} A grouping or keys column giving the forecast hierarchy.
\item \code{city} A grouping or keys column giving the forecast hierarchy.
\item \code{widgets} The outcome being forecasted.
}
}
}
\usage{
widgets
}
\description{
The output of a hierarchical forecast in a keys, index, outcome format.
}
\keyword{datasets}
